+++ packages/cheqd-blockchain-modules/src/common/index.js
@@ -1,6 +1,5 @@
 export { default as createInternalCheqdModule } from './create-internal-cheqd-module';
-export { default as injectCheqd } from './inject-cheqd';
-export { default as CheqdApiProvider } from './cheqd-api-provider';
+export { default as withCheqd } from './with-cheqd';
 export { default as injectParams } from './inject-params';
 export { default as injectPublicKeys } from './inject-public-keys';
 export { default as withParams } from './with-params';