+++ packages/cheqd-blockchain-modules/src/offchain-signatures/module.js
@@ -1,10 +1,10 @@
 import { AbstractOffchainSignaturesModule } from '@docknetwork/credential-sdk/modules/abstract';
 import { OffchainSignatureParams } from '@docknetwork/credential-sdk/types';
-import { injectCheqd, withParams } from '../common';
+import { withCheqd, withParams } from '../common';
 import CheqdOffchainSignaturesInternalModule from './internal';
 
 export default class CheqdOffchainSignaturesModule extends withParams(
-  injectCheqd(AbstractOffchainSignaturesModule),
+  withCheqd(AbstractOffchainSignaturesModule),
 ) {
   static CheqdOnly = class extends CheqdOffchainSignaturesInternalModule {
     static Params = OffchainSignatureParams;